requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
Django==5.0.5
django-bootstrap4==24.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==23.0.0
lxml==5.3.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyxlsb==1.0.10
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
xlrd==2.0.1
